US i94 immigrant travel datawarehouse
This project receives and prepares I94 Immigrant data from the United States for analysis. This data represents every immigrant departure at every port by date. It also has other pertinent information including the address the immigrant stayed in the US as well as whether the records was matched to an arrival i94 or not. The project intends to include dimensions of state-of-residence demographics and average monthly temps at departing port.